Blog/requirements.txt
flask
python-dotenv
google-auth-oauthlib
google-auth
google-api-python-client
python-docx
requests
mysqlclient
